+++ examples/codeSample/synchronizationExample.js
@@ -1,0 +1,61 @@
+let MetaApi = require('metaapi.cloud-sdk').default;
+
+let token = process.env.TOKEN || '<put in your token here>';
+let accountId = process.env.ACCOUNT_ID || '<put in your account id here>';
+const api = new MetaApi(token);
+
+async function testMetaApiSynchronization() {
+  try {
+    const account = await api.metatraderAccountApi.getAccount(accountId);
+    // wait until account is deployed and connected to broker
+    console.log('Deploying account');
+    await account.deploy();
+    console.log('Waiting for API server to connect to broker (may take couple of minutes)');
+    await account.waitConnected();
+
+    // connect to MetaApi API
+    let connection = await account.connect();
+
+    // wait until terminal state synchronized to the local state
+    console.log('Waiting for SDK to synchronize to terminal state (may take some time depending on your history size)');
+    await connection.waitSynchronized();
+
+    // access local copy of terminal state
+    console.log('Testing terminal state access');
+    let terminalState = connection.terminalState;
+    console.log('connected:', terminalState.connected);
+    console.log('connected to broker:', terminalState.connectedToBroker);
+    console.log('account information:', terminalState.accountInformation);
+    console.log('positions:', terminalState.positions);
+    console.log('orders:', terminalState.orders);
+    console.log('specifications:', terminalState.specifications);
+    console.log('EURUSD specification:', terminalState.specification('EURUSD'));
+    console.log('EURUSD price:', terminalState.price('EURUSD'));
+
+    // trade
+    console.log('Submitting pending order');
+    try {
+      let result = await
+      connection.createLimitBuyOrder('GBPUSD', 0.07, 1.0, 0.9, 2.0, {
+        comment: 'comm',
+        clientId: 'TE_GBPUSD_7hyINWqAlE'
+      });
+      console.log('Trade successful, result code is ' + result.stringCode);
+    } catch (err) {
+      console.log('Trade failed with result code ' + err.stringCode);
+    }
+
+    // finally, undeploy account after the test
+    console.log('Undeploying account so that it does not consume any unwanted resources');
+    await account.undeploy();
+
+    process.exit();
+  } catch (err) {
+    console.error(err);
+  }
+}
+
+testMetaApiSynchronization();
+
+
+