+++ changelog.md
@@ -9,6 +9,7 @@
   - added the ability to work in web apps
   - added the ability to retrieve Metatrader account by account access token
   - added the verification access depending on the token to API
+  - added websocket and http client timeouts
 
 5.0.2
   - minor bugfix